Refactored Models and changed seed data
